initial_data_pipeline/requirements.txt
requests
passlib
pandas==2.0.3
numpy
scikit-surprise
scikit-learn==1.3.2
python-dotenv